Extract out SearchBar from NPO management
