+++ src/pages/npo-management/index.tsx
@@ -6,6 +6,7 @@ import NpoModal from "./npo-modal";
 import { useEffect, useState } from "react";
 import { api } from "~/utils/api";
 import { useRouter } from "next/router";
+import SearchBar from "../../components/SearchBar";
 
 const NpoManagement = () => {
   const ENTRIES_PER_PAGE = 10;
@@ -237,47 +238,5 @@ const ViewOppButton = ({ onClick }: { onClick: () => void }) => {
       </svg>
       View Events
     </button>
-  );
-};
-
-const SearchBar = ({
-  searchTerm,
-  setSearchTerm,
-}: {
-  searchTerm: string;
-  setSearchTerm: (searchTerm: string) => void;
-}) => {
-  return (
-    <div className="mb-[40px] w-[90vw] md:w-[50vw] md:max-w-xl">
-      <label className="sr-only mb-2 text-sm font-medium text-gray-900 dark:text-white">
-        Search
-      </label>
-      <div className="relative">
-        <div className="pointer-events-none absolute inset-y-0 start-0 flex items-center ps-3">
-          <svg
-            className="h-4 w-4 text-gray-500 dark:text-gray-400"
-            aria-hidden="true"
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 20 20"
-          >
-            <path
-              stroke="currentColor"
-              stroke-linecap="round"
-              stroke-linejoin="round"
-              stroke-width="2"
-              d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
-            />
-          </svg>
-        </div>
-        <input
-          value={searchTerm}
-          onChange={(e) => setSearchTerm(e.target.value)}
-          className="block w-full rounded-lg border border-gray-300 bg-gray-50 p-4 ps-10 text-sm text-gray-900 focus:border-blue-500 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
-          placeholder="Search..."
-          required
-        />
-      </div>
-    </div>
   );
 };
